backend/requirements.txt
fastapi
uvicorn[standard]
SQLAlchemy
pydantic
python-multipart
Jinja2
python-jose[cryptography]
passlib==1.7.4
bcrypt==3.2.0
tenacity
WeasyPrint
psycopg2-binary
python-dotenv
